+++ src/main/java/za/co/hpsc/web/constants/SystemConstants.java
@@ -21,7 +21,6 @@ public class SystemConstants {
 
     public static final String ISO_DATE_FORMAT = "yyyy-MM-dd";
     public static final String ISO_DATE_TIME_FORMAT = ISO_DATE_FORMAT + " HH:mm";
-    public static final String T_SEPARATED_DATE_TIME_FORMAT = "yyyy-MM-dd'T'HH:mm:ss";
     public static final String LONG_DATE_FORMAT = "dd MMMM yyyy";
     public static final String LONG_DATE_TIME_FORMAT = LONG_DATE_FORMAT + " HH:mm";
 }